Create and destroy CUFFT plan for each data buffer.
We used to create the CUFFT plan once (assuming no
dynamic change of the coarse channel count) and destroy
it at exit time.  This was nice and simple but meant that the
GPU memory footprint for CUFFT remained in place at all
times.  This footprint can be substantial (ex, for a 0.5 gigasample
real input, the CUFFT plan consumed 2GB of GPU memory).

With this change, we execute the plan right before calling cufft
and destroy it right after.   This frees up memory at a slight cost
in per instance performance.