Fixed bug which made fine channels wrong by a factor of two.
In the making of the strided iterator used in the fluffing
of 8bit raw data to cuda floats we had hard coded the number
of polarizations as 2.

This is bad in principle and, in the case of FAST, wrong in fact
because the number of pols per data stream is 1!

So now we use the passed-in N_POLS_PER_BEAM.