Comment out coarse channel averaging for FAST.
Not that useful when we do not have coarse channelized
data (ie, one coarse channel) and not doing it saves some
needed time.

Also a couple of debug things and some untested (and commented
out) code variations to do the cuFFT in place.  This is not needed
at this time.