Pass source ID (and thus pol number) on to the output thread.
In non-FAST versions of serendip6 both pols were present in
the input data and were distinguished by position in the input
data buffer.

For the FAST version only one pol is present. Which pol it is
must be explicitly communicated to the output thread.

This is a kind of ugly hack but I wanted to get it in. I may make
it nicer later.

This change also includes making both the beam and pol numbers
part of the output FITS file name.