Support for multicasting.
Added support for multicasting. Only the FAST
installation uses this for now. It is turned
off (via a #define) in the committed source.

Method:
We are unable to join a multicast group directly via the packet
socket because the "socket option level" (IPPROTO_IP) is not
valid for packet sockets. So we open another, generic, socket
at the same IP and port and use this "associate" socket to join
the multicast group.