Mod the use of status shmem for threshold and nhits.
SNR threshold will come from status shared memory if there.
Otherwise the compiled in default will be used.  This allows
the threshold to be specified on the hashpipe command line as:
	-o POWTHRSH=N
where N in the desired threshold.

The number of hits for the most recently completed spectra is
now stored in status shared memory as NUMHITS.