Remove unused transpose data structure allocation.
A cleanup mod.

In the very early development of serendip6 there was some thought given to transposing
the input data (which arrives all coarse channels for one time followed by all coarse channels
for the next time...) to a format more natural for the FFT (all times for one coarse channel
followed by all times for the next coarse channel...). But then we just used striding in cufft to
achieve this.