Mods to allow cub allocator integration.
We need -std=c++11 for gcc as well as for nvcc.

The headers needed by cub were moved to s6GPU.h (from s6GPU.cu)
so that we do not have to explicitly include these in every file that
includes s6GPU.h.  Another option would have been to move the device
vector declarations to a new header file that would then be included in
s6GPU.cu.  But this way seems simpler.