Drop the R2C FFT "padding" element from down stream processing.
The length of the (complex) output vector for a real to complex cuFFT is required
to be n/2+1 where n is the length of the (real) input vector.  The +1 was causing
more-than-needed allocation from the cub allocator for down stream processing.
So we drop the padding element at the time we compute the power spectrum.