Make cudaThreadSynchronize() easy to turn off/on (and turn it off).
I don't think we need this.  All kernel launches are in one stream
by default and tasks within a stream run sequentially. But, if results
get wonky, we may return to this.

The reason to turn it off is that this may speed things up a very small
amount (the CPU code continues immediately after the thrust call).

At any rate, it is easy to toggle.

Note: cudaThreadSynchronize() is now deprecated by NVidia in
favor of cudaDeviceSynchronize().