Add a REALLOC_NONE version of spectroscopy().
Added a version of spectroscopy() that does no(*) GPU dynamic memory allocation.

One of the three reallocation policies are now selected by, in s6GPU.h:
// uncomment exactly one REALLOC_x
//#define REALLOC_STD
//#define REALLOC_NONE

This committed version still uses REALLOC_CUB.

The reason for adding a REALLOC_NONE version is to test, and perhaps
use, a reduction (by half) of frequency resolution in order to regain the
ability to run on both pols without packet loss.  There is no packet loss
with one pol, regardless of reallocation policy.  But with both REALLOC_CUB
and REALLOC_STD there is packet loss with two pols (memory reallocation
is expensive).

The reduction in frequency resolution is not part of this commit.

(*) Actually, we had to go with minimal, rather than no, reallocation.  Tests
with a reduction in frequency resolution still used too much GPU memory
with two pols.  In this commit we reallocate the cuFFT plan and the GPU
input vector.