First commit after more than a year away from the code.
All of the mods here are built into the current production executables.

This includes changes to timestamp data (hits) blocks at sub second
resolution.  This is necessary because the integration time is less than
a second.  However, this higher resolution is not yet being written to
the output FITS files.